+++ mofumofu-server/crates/mofumofu-entity/src/user_roles.rs
@@ -37,4 +37,10 @@ pub enum Relation {
     Granter,
 }
 
+impl Related<UsersEntity> for Entity {
+    fn to() -> RelationDef {
+        Relation::User.def()
+    }
+}
+
 impl ActiveModelBehavior for ActiveModel {}